checkGitDir
