+++ lib/cache.js
@@ -85,6 +85,7 @@ var mkdir = require("mkdirp")
   , isGitUrl = require("./utils/is-git-url.js")
   , pathIsInside = require("path-is-inside")
   , getCacheStat = require("npm-cache-dir").getCacheStat
+  , checkGitDir = require("./git-cache.js").checkGitDir
 
 cache.usage = "npm cache add <tarball file>"
             + "\nnpm cache add <folder>"
@@ -445,54 +446,14 @@ function addRemoteGit (u, parsed, name, silent, cb_) {
     log.verbose("addRemoteGit", [u, co])
 
     var p = path.join(npm.config.get("cache"), "_git-remotes", v)
-
-    checkGitDir(p, u, co, origUrl, silent, function(er, data) {
-      chmodr(p, npm.modes.file, function(erChmod) {
+      , g = npm.config.get("git")
+      , t = npm.tmp
+      , ca = npm.cache
+    checkGitDir(p, u, co, origUrl, silent, ca, t, g, log, gitDoneCb
+    , function(er, data) {
+      chmodr(p, npm.modes.file, function (erChmod) {
         if (er) return cb(er, data)
         return cb(erChmod, data)
-      })
-    })
-  })
-}
-
-var cloneGitRemote = require("./git-cache.js").cloneGitRemote
-function checkGitDir (p, u, co, origUrl, silent, cb) {
-  // xxx git-module refactor
-  var g = npm.config.get("git")
-    , t = npm.tmp
-    , c = npm.cache
-
-  fs.stat(p, function (er, s) {
-    if (er) return cloneGitRemote(p, u, co, origUrl, silent
-      , c, t, g, log, gitDoneCb, cb)
-    if (!s.isDirectory()) return rm(p, function (er){
-      if (er) return cb(er)
-      cloneGitRemote(p, u, co, origUrl, silent, c, t, g, log, gitDoneCb, cb)
-    })
-
-    var git = npm.config.get("git")
-    var args = [ "config", "--get", "remote.origin.url" ]
-    var env = gitEnv()
-
-    // check for git
-    which(git, function (err) {
-      if (err) {
-        err.code = "ENOGIT"
-        return cb(err)
-      }
-      exec(git, args, {cwd: p, env: env}, function (er, stdout, stderr) {
-        stdoutTrimmed = (stdout + "\n" + stderr).trim()
-        if (er || u !== stdout.trim()) {
-          log.warn( "`git config --get remote.origin.url` returned "
-                  + "wrong result ("+u+")", stdoutTrimmed )
-          return rm(p, function (er) {
-            if (er) return cb(er)
-            cloneGitRemote(p, u, co, origUrl, silent
-              , c, t, g, log, gitDoneCb, cb)
-          })
-        }
-        log.verbose("git remote.origin.url", stdoutTrimmed)
-        archiveGitRemote(p, u, co, origUrl, cb)
       })
     })
   })
@@ -504,29 +465,6 @@ function gitDoneCb (tmp, resolved, cb) {
     if (data) data._resolved = resolved
     cb(er, data)
   })
-}
-
-function archiveGitRemote (p, u, co, origUrl, cb) {
-  // xxx git-module refactor
-  var g = npm.config.get("git")
-    , t = npm.tmp
-    , c = npm.cache
-
-  require("./git-cache.js")
-    .archiveGitRemote(p, u, co, origUrl, c, t, g, log, gitDoneCb, cb)
-}
-
-var gitEnv_
-function gitEnv () {
-  // git responds to env vars in some weird ways in post-receive hooks
-  // so don't carry those along.
-  if (gitEnv_) return gitEnv_
-  gitEnv_ = {}
-  for (var k in process.env) {
-    if (!~['GIT_PROXY_COMMAND','GIT_SSH','GIT_SSL_NO_VERIFY'].indexOf(k) && k.match(/^GIT/)) continue
-    gitEnv_[k] = process.env[k]
-  }
-  return gitEnv_
 }
