+++ lib/git-cache.js
@@ -10,6 +10,50 @@ var chownr = require("chownr")
   , spawn = require("child_process").spawn
   , which = require("which")
 
+
+// 1. checkGitDir(cacheDir) ? 4. : 3. (rm cacheDir if necessary)
+// 2. git clone --mirror u cacheDir
+// 3. cd cacheDir && git fetch -a origin
+// 4. git archive /tmp/random.tgz
+// 5. addLocalTarball(/tmp/random.tgz) <gitref> --format=tar --prefix=package/
+// silent flag is used if this should error quietly
+exports.checkGitDir = checkGitDir
+function checkGitDir (p, u, co, origUrl, silent, ca, t, git, log, gitDoneCb, cb) {
+  fs.stat(p, function (er, s) {
+    if (er) return cloneGitRemote(p, u, co, origUrl, silent
+      , ca, t, git, log, gitDoneCb, cb)
+    if (!s.isDirectory()) return rm(p, function (er){
+      if (er) return cb(er)
+      cloneGitRemote(p, u, co, origUrl, silent, ca, t, git, log, gitDoneCb, cb)
+    })
+
+    var args = [ "config", "--get", "remote.origin.url" ]
+    var env = gitEnv()
+
+    // check for git
+    which(git, function (err) {
+      if (err) {
+        err.code = "ENOGIT"
+        return cb(err)
+      }
+      exec(git, args, {cwd: p, env: env}, function (er, stdout, stderr) {
+        stdoutTrimmed = (stdout + "\n" + stderr).trim()
+        if (er || u !== stdout.trim()) {
+          log.warn( "`git config --get remote.origin.url` returned "
+                  + "wrong result ("+u+")", stdoutTrimmed )
+          return rm(p, function (er) {
+            if (er) return cb(er)
+            cloneGitRemote(p, u, co, origUrl, silent
+              , c, t, git, log, gitDoneCb, cb)
+          })
+        }
+        log.verbose("git remote.origin.url", stdoutTrimmed)
+        archiveGitRemote(p, u, co, origUrl, ca, t, git, log, gitDoneCb, cb)
+      })
+    })
+  })
+}
+
 exports.cloneGitRemote = cloneGitRemote
 function cloneGitRemote (p, u, co, origUrl, silent, ca, t, git, log, gitDoneCb, cb) {
   mkdir(p, function (er) {
@@ -129,8 +173,6 @@ function archiveGitRemote (p, u, co, origUrl, ca, t, git, log, gitDoneCb, cb) {
     })
   }
 }
-
-
 
 var gitEnv_
 function gitEnv () {
